requirements.txt
fastapi
uvicorn
gtts
python-multipart
python-dotenv